update blade and sould plugin
