+++ sigma/plugins/games/blade_and_souls/plugin.yml
@@ -1,0 +1,31 @@
+name: Blade and Souls
+
+categories:
+  - games
+
+enabled: true
+
+commands:
+  - name:    bnsstats
+    global:  true
+    enabled: true
+    sfw:     true
+    usage:   "{pfx:s}{cmd:s} EU Lucia Konohana"
+    description:
+      The Blade and Soul module provides in-depth character information.
+
+  - name:    bnsatt
+    global:  true
+    enabled: true
+    sfw:     true
+    usage:   "{pfx:s}{cmd:s} EU Lucia Konohana"
+    description:
+      Provides highly detailed Attack Statistics
+
+  - name:    bnsdef
+    global:  true
+    enabled: true
+    sfw:     true
+    usage:   "{pfx:s}{cmd:s} EU Lucia Konohana"
+    description:
+      Provides highly detailed Defense Statistics
